+++ java/hsfs/src/main/java/com/logicalclocks/hsfs/metadata/FeatureGroupApi.java
@@ -24,7 +24,6 @@ import com.logicalclocks.hsfs.FeatureGroupBase;
 import com.logicalclocks.hsfs.FeatureGroupCommit;
 import com.logicalclocks.hsfs.FeatureStoreBase;
 import com.logicalclocks.hsfs.FeatureStoreException;
-import com.logicalclocks.hsfs.IngestionRun;
 import com.logicalclocks.hsfs.JobConfiguration;
 import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;
@@ -298,48 +297,6 @@ public class FeatureGroupApi {
     }
 
     return featureGroup;
-  }
-
-  public void saveIngestionRun(FeatureGroupBase featureGroupBase, IngestionRun ingestionRun)
-      throws FeatureStoreException, IOException {
-    HopsworksClient hopsworksClient = HopsworksClient.getInstance();
-    String pathTemplate = HopsworksClient.PROJECT_PATH
-        + FeatureStoreApi.FEATURE_STORE_PATH
-        + FEATURE_GROUP_INGESTION_RUN;
-
-    String uri = UriTemplate.fromTemplate(pathTemplate)
-        .set("projectId", hopsworksClient.getProject().getProjectId())
-        .set("fsId", featureGroupBase.getFeatureStore().getId())
-        .set("fgId", featureGroupBase.getId())
-        .expand();
-
-    HttpPost postRequest = new HttpPost(uri);
-    postRequest.setEntity(hopsworksClient.buildStringEntity(ingestionRun));
-    hopsworksClient.handleRequest(postRequest);
-  }
-
-  public IngestionRun getIngestionRun(FeatureGroupBase featureGroupBase,  Map<String, String> queryParams)
-      throws IOException, FeatureStoreException {
-    HopsworksClient hopsworksClient = HopsworksClient.getInstance();
-    String pathTemplate = HopsworksClient.PROJECT_PATH
-        + FeatureStoreApi.FEATURE_STORE_PATH
-        + FEATURE_GROUP_INGESTION_RUN;
-
-    UriTemplate uriTemplate = UriTemplate.fromTemplate(pathTemplate)
-        .set("projectId", hopsworksClient.getProject().getProjectId())
-        .set("fsId", featureGroupBase.getFeatureStore().getId())
-        .set("fgId", featureGroupBase.getId());
-      
-    if (queryParams != null) {
-      for (Map.Entry<String, String> entry: queryParams.entrySet()) {
-        uriTemplate.set(entry.getKey(), entry.getValue());
-      }
-    }
-
-    String uri = uriTemplate.expand();
-
-    LOGGER.info("Sending metadata request: " + uri);
-    return hopsworksClient.handleRequest(new HttpGet(uri), IngestionRun.class);
   }
 
   private <T extends FeatureGroupBase> void checkFeatures(T fg) {
