+++ java/spark/src/main/java/com/logicalclocks/hsfs/spark/engine/SparkEngine.java
@@ -27,6 +27,7 @@ import com.logicalclocks.hsfs.engine.EngineBase;
 import com.logicalclocks.hsfs.metadata.DatasetApi;
 import com.logicalclocks.hsfs.metadata.FeatureGroupApi;
 import com.logicalclocks.hsfs.metadata.HopsworksExternalClient;
+import com.logicalclocks.hsfs.metadata.IngestionRunApi;
 import com.logicalclocks.hsfs.spark.constructor.Query;
 import com.logicalclocks.hsfs.spark.engine.hudi.HudiEngine;
 import com.logicalclocks.hsfs.DataFormat;
@@ -131,6 +132,7 @@ public class SparkEngine extends EngineBase {
   private final StorageConnectorUtils storageConnectorUtils = new StorageConnectorUtils();
   private FeatureGroupUtils featureGroupUtils = new FeatureGroupUtils();
   protected FeatureGroupApi featureGroupApi = new FeatureGroupApi();
+  protected IngestionRunApi ingestionRunApi = new IngestionRunApi();
   private final KafkaEngine kafkaEngine;
 
   private static SparkEngine INSTANCE = null;
@@ -574,7 +576,7 @@ public class SparkEngine extends EngineBase {
 
     String endingCheckPoint = kafkaEngine.kafkaGetOffsets(featureGroupBase, writeOptions, true);
 
-    featureGroupApi.saveIngestionRun(featureGroupBase, new IngestionRun(startingCheckPoint, endingCheckPoint));
+    ingestionRunApi.saveIngestionRun(featureGroupBase, new IngestionRun(startingCheckPoint, endingCheckPoint));
   }
 
   public <S> StreamingQuery writeStreamDataframe(FeatureGroupBase featureGroupBase, Dataset<Row> dataset,
@@ -635,7 +637,7 @@ public class SparkEngine extends EngineBase {
         public void onQueryTerminated(QueryTerminatedEvent queryTerminated) {
           try {
             String endingCheckPoint = kafkaEngine.kafkaGetOffsets(featureGroupBase, writeOptions, true);
-            featureGroupApi.saveIngestionRun(featureGroupBase, new IngestionRun(startingCheckPoint, endingCheckPoint));
+            ingestionRunApi.saveIngestionRun(featureGroupBase, new IngestionRun(startingCheckPoint, endingCheckPoint));
           } catch (Exception e) {
             e.printStackTrace();
           }
